refactor play button
