+++ components/player/Player.tsx
@@ -5,8 +5,7 @@ import { zEpisodeData } from "@/types/podcastTypes";
 import React, { useRef } from "react";
 import { z } from "zod";
 import { create } from "zustand";
-import { Button } from "../ui/button";
-import { PauseIcon, PlayIcon } from "@radix-ui/react-icons";
+import PlayButton from "./PlayButton";
 
 type Props = {
   className?: string;
@@ -35,28 +34,6 @@ const Player = ({ className }: Props) => {
   const [isPlaying, setIsPlaying] = React.useState(false);
   const episodeData = AudioStore((state) => state.episodeData);
 
-  const handlePlayButtonClick = () => {
-    if (!isPlaying) {
-      if (episodeData.audio_url === "" || episodeData.audio_url === undefined)
-        return;
-      play();
-    } else {
-      pause();
-    }
-  };
-
-  const play = () => {
-    void player.current?.play();
-  };
-
-  const pause = () => {
-    void player.current?.pause();
-  };
-
-  const playerSeek = (time: number) => {
-    if (player.current) player.current.currentTime = time;
-  };
-
   return (
     <nav className={cn("h-20 border-t-2 border-border", className)}>
       <audio
@@ -71,23 +48,10 @@ const Player = ({ className }: Props) => {
         // onRateChange={(e) => refreshAudioData(e)}
       />
       <div className="flex h-full w-full items-center justify-center">
-        <PlayButton isPlaying={isPlaying} handleClick={handlePlayButtonClick} />
+        <PlayButton isPlaying={isPlaying} playerRef={player} />
       </div>
     </nav>
   );
 };
 
 export default Player;
-
-type PlayButtonProps = {
-  isPlaying: boolean;
-  handleClick: () => void;
-};
-
-const PlayButton = ({ isPlaying, handleClick }: PlayButtonProps) => {
-  return (
-    <Button size="roundedIcon" onClick={handleClick}>
-      {isPlaying ? <PauseIcon /> : <PlayIcon />}
-    </Button>
-  );
-};
